back to float

+++ matrix_vector_mul.c
@@ -2,9 +2,9 @@
 #include <stdlib.h>
 #include <time.h>
 
-void matrix_vector_product(long long int *a, long long int *v, long long int *res, int matrix_size) {
+void matrix_vector_product(long double *a, long double *v, long double *res, int matrix_size) {
     for (int i = 0; i < matrix_size; i++) {
-        long long int sum = 0;
+        long double sum = 0;
         for (int j = 0; j < matrix_size; j++) {
             sum += a[i * matrix_size + j] * v[j];
         } 
@@ -13,27 +13,19 @@ void matrix_vector_product(long long int *a, long long int *v, long long int *re
 }
 
 int main(int argc, char **argv) {
-    long long int *a, *v, *res; 
+    long double *a, *v, *res; 
     int matrix_size = atoi(argv[1]);
     
-    a = (long long int *) malloc(matrix_size * matrix_size * sizeof(long long int));
-    v = (long long int *) malloc(matrix_size * sizeof(long long int));
-    res = (long long int *) malloc(matrix_size * sizeof(long long int));
+    a = (long double *) malloc(matrix_size * matrix_size * sizeof(long double));
+    v = (long double *) malloc(matrix_size * sizeof(long double));
+    res = (long double *) malloc(matrix_size * sizeof(long double));
     
-    // printf("a: ");
     for (int i = 0; i < matrix_size; i++) {
         for (int j = 0; j < matrix_size; j++) {
-            a[i * matrix_size + j] = (long long int) i * matrix_size + j;
-            // printf("%lld ", a[i * matrix_size + j]);
+            a[i * matrix_size + j] = (long double) i * matrix_size + j;
         }
-        // printf("\n");
-        v[i] = (long long int) i;
+        v[i] = (long double) i;
     }
-    // printf("v: ");
-    // for (int i = 0; i < matrix_size; i++) {
-    //     printf("%lld ", v[i]);
-    // }
-    // printf("\n");
 
     clock_t st, en;
     st = clock();
@@ -41,7 +33,7 @@ int main(int argc, char **argv) {
     en = clock();
 
     for (int i = 0; i < matrix_size; i++) {
-        printf("%lld\n", res[i]);
+        printf("%.2Lf\n", res[i]);
     }
     
     printf("Time taken by CPU: %f seconds\n", (double)(en - st) / CLOCKS_PER_SEC);
